Added SidecarWriter
