+++ osxphotos/cli/exiftool_cli.py
@@ -11,10 +11,12 @@ from osxphotos import PhotosDB
 from osxphotos._constants import OSXPHOTOS_EXPORT_DB
 from osxphotos._version import __version__
 from osxphotos.configoptions import ConfigOptions, ConfigOptionsLoadError
+from osxphotos.exifwriter import exiftool_json_sidecar
 from osxphotos.export_db import ExportDB, ExportDBInMemory
 from osxphotos.export_db_utils import export_db_get_config
+from osxphotos.exportoptions import ExportOptions, ExportResults
 from osxphotos.fileutil import FileUtil, FileUtilNoOp
-from osxphotos.photoexporter import ExportOptions, ExportResults, PhotoExporter
+from osxphotos.photoexporter import PhotoExporter
 from osxphotos.utils import pluralize
 
 from .cli_params import DB_OPTION, THEME_OPTION, TIMESTAMP_OPTION, VERBOSE_OPTION
@@ -382,7 +384,7 @@ def process_files(
             with export_db.get_file_record(file) as rec:
                 rec.dest_sig = fileutil.file_sig(file)
                 rec.export_options = export_options.bit_flags
-                rec.exifdata = exporter.exiftool_json_sidecar(export_options)
+                rec.exifdata = exiftool_json_sidecar(photo, export_options)
 
             report_writer.write(results)
             count += 1
